requirements.txt
altgraph==0.17.4
attrs==24.2.0
blinker==1.8.2
bottle==0.13.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et_xmlfile==2.0.0
Flask==3.0.3
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.1.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
plotly==5.24.1
proxy-tools==0.1.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pythonnet==3.0.4
pytz==2024.2
pywebview==5.3.2
pywin32-ctypes==0.2.3
requests==2.32.3
retrying==1.3.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
